requirements.txt
appdirs==1.4.4
asgiref==3.3.4
astroid==2.7.2
backcall==0.2.0
backports.entry-points-selectable==1.1.0
black==21.7b0
certifi==2021.5.30
click==8.0.1
coverage==5.5
decorator==5.0.9
distlib==0.3.2
Django==3.2
django-environ==0.4.5
django-extensions==3.1.3
djangorestframework==3.12.4
Faker==8.12.1
filelock==3.0.12
ipdb==0.13.9
ipython==7.27.0
ipython-genutils==0.2.0
isort==5.9.3
jedi==0.18.0
lazy-object-proxy==1.6.0
matplotlib-inline==0.1.2
mccabe==0.6.1
mypy-extensions==0.4.3
parso==0.8.2
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
pipenv==2021.5.29
platformdirs==2.2.0
prompt-toolkit==3.0.20
psycopg2-binary==2.9.1
ptyprocess==0.7.0
Pygments==2.10.0
pylint==2.10.2
python-dateutil==2.8.2
python-decouple==3.4
python-dotenv==0.19.0
pytz==2021.1
regex==2021.8.28
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
tomli==1.2.1
traitlets==5.0.5
virtualenv==20.7.2
virtualenv-clone==0.5.6
wcwidth==0.2.5
wrapt==1.12.1
